+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * [#301](https://github.com/tboox/xmake/issues/301): Improve precompiled header file
 * [#322](https://github.com/tboox/xmake/issues/322): Add `option.add_features`, `option.add_cxxsnippets` and `option.add_csnippets`
 * Remove some deprecated interfaces of xmake 1.x, e.g. `add_option_xxx`
+* [#327](https://github.com/tboox/xmake/issues/327): Support conan package manager for `lib.detect.find_package` 
 
 ### Bugs fixed
 
@@ -558,6 +559,7 @@
 * [#301](https://github.com/tboox/xmake/issues/301): 改进编译预处理头文件以及依赖头文件生成，编译速度提升30%
 * [#322](https://github.com/tboox/xmake/issues/322): 添加`option.add_features`, `option.add_cxxsnippets` 和 `option.add_csnippets`
 * 移除xmake 1.x的一些废弃接口, 例如：`add_option_xxx`
+* [#327](https://github.com/tboox/xmake/issues/327): 改进`lib.detect.find_package`增加对conan包管理器的支持
 
 ### Bugs修复